CDISCFilteredDataset R6 class

CDISCFilteredDataset R6 class

CDISCFilteredDataset R6 class

Super classes

teal.slice::FilteredDataset -> teal.slice::DefaultFilteredDataset -> CDISCFilteredDataset

Methods

Public methods

CDISCFilteredDataset$get_call()

CDISCFilteredDataset$get_filter_overview_info()

CDISCFilteredDataset$clone()

Inherited methods

teal.slice::FilteredDataset$add_to_eval_env()

teal.slice::FilteredDataset$get_data()

teal.slice::FilteredDataset$get_data_reactive()

teal.slice::FilteredDataset$get_dataname()

teal.slice::FilteredDataset$get_dataset()

teal.slice::FilteredDataset$get_filter_states()

teal.slice::FilteredDataset$get_filterable_varnames()

teal.slice::FilteredDataset$get_filtered_dataname()

teal.slice::FilteredDataset$get_hash()

teal.slice::FilteredDataset$get_join_keys()

teal.slice::FilteredDataset$get_keys()

teal.slice::FilteredDataset$get_metadata()

teal.slice::FilteredDataset$get_varlabels()

teal.slice::FilteredDataset$get_varnames()

teal.slice::FilteredDataset$queues_empty()

teal.slice::FilteredDataset$server()

teal.slice::FilteredDataset$ui()

teal.slice::DefaultFilteredDataset$get_filter_state()

teal.slice::DefaultFilteredDataset$initialize()

teal.slice::DefaultFilteredDataset$remove_filter_state()

teal.slice::DefaultFilteredDataset$set_filter_state()

teal.slice::DefaultFilteredDataset$srv_add_filter_state()

teal.slice::DefaultFilteredDataset$ui_add_filter_state()

Method get_call()

Get filter expression

This functions returns filter calls equivalent to selected items within each of filter_states. Configuration of the calls is constant and depends on filter_states type and order which are set during initialization. This class contains single FilterStates which contains single ReactiveQueue and all FilterState objects applies to one argument (...) in dplyr::filter call. It's also possible within this class to return merge call with other data.frame which is a parent.

Usage

CDISCFilteredDataset$get_call()

Returns

filter call or list of filter calls

Method get_filter_overview_info()

Get filter overview rows of a dataset

Usage

CDISCFilteredDataset$get_filter_overview_info()

Returns

(matrix) matrix of observations and subjects

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCFilteredDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.