FilterState object for factor or character variable

FilterState object for factor or character variable

FilterState object for factor or character variable

Note

Casts the passed object to character before validating the input making it possible to pass any object coercible to character to this method.

Super class

teal.slice::FilterState -> ChoicesFilterState

Methods

Public methods

ChoicesFilterState$new()

ChoicesFilterState$is_any_filtered()

ChoicesFilterState$get_call()

ChoicesFilterState$set_state()

ChoicesFilterState$set_state_reactive()

ChoicesFilterState$set_selected()

ChoicesFilterState$clone()

Inherited methods

teal.slice::FilterState$destroy_observers()

teal.slice::FilterState$format()

teal.slice::FilterState$get_dataname()

teal.slice::FilterState$get_keep_na()

teal.slice::FilterState$get_selected()

teal.slice::FilterState$get_state()

teal.slice::FilterState$get_varlabel()

teal.slice::FilterState$get_varname()

teal.slice::FilterState$print()

teal.slice::FilterState$server()

teal.slice::FilterState$set_keep_na()

teal.slice::FilterState$set_keep_na_reactive()

teal.slice::FilterState$set_na_rm()

teal.slice::FilterState$set_selected_reactive()

teal.slice::FilterState$ui()

Method new()

Initialize a FilterState object

Usage

ChoicesFilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(character or factor)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method is_any_filtered()

Answers the question of whether the current settings and values selected actually filters out any values.

Usage

ChoicesFilterState$is_any_filtered()

Returns

logical scalar

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> %in% c(<values selected>) with optional is.na(<varname>).

Usage

ChoicesFilterState$get_call()

Returns

(call)

Method set_state()

Set state

Usage

ChoicesFilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

Method set_state_reactive()

Set state when using set_filter_state

Usage

ChoicesFilterState$set_state_reactive(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

Method set_selected()

Sets the selected values of this ChoicesFilterState.

Usage

ChoicesFilterState$set_selected(value)

Arguments

value

(character) the array of the selected choices. Must not contain NA values.

Returns

invisibly NULL

Examples

filter <- teal.slice:::ChoicesFilterState$new(c("a", "b", "c"), varname = "name")
filter$set_selected(c("c", "a"))

Method clone()

The objects of this class are cloneable with this method.

Usage

ChoicesFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal.slice:::ChoicesFilterState$new(
  c(LETTERS, NA),
  varname = "x",
  input_dataname = as.name("data"),
  extract_type = character(0)
)
isolate(filter_state$get_call())
#> x %in% c("A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", 
#>     "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", 
#>     "X", "Y", "Z")
isolate(filter_state$set_selected("B"))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$get_call())
#> is.na(x) | x == "B"

## ------------------------------------------------
## Method `ChoicesFilterState$set_selected`
## ------------------------------------------------

filter <- teal.slice:::ChoicesFilterState$new(c("a", "b", "c"), varname = "name")
filter$set_selected(c("c", "a"))