FilterState object for POSIXct variable

Manages choosing a range of date-times

Note

Casts the passed object to POSIXct before validating the input making it possible to pass any object coercible to POSIXct to this method.

Super class

teal.slice::FilterState -> DatetimeFilterState

Methods

Public methods

DatetimeFilterState$new()

DatetimeFilterState$format()

DatetimeFilterState$is_any_filtered()

DatetimeFilterState$get_call()

DatetimeFilterState$set_selected()

DatetimeFilterState$clone()

Inherited methods

teal.slice::FilterState$destroy_observers()

teal.slice::FilterState$get_dataname()

teal.slice::FilterState$get_keep_na()

teal.slice::FilterState$get_selected()

teal.slice::FilterState$get_state()

teal.slice::FilterState$get_varlabel()

teal.slice::FilterState$get_varname()

teal.slice::FilterState$print()

teal.slice::FilterState$server()

teal.slice::FilterState$set_keep_na()

teal.slice::FilterState$set_na_rm()

teal.slice::FilterState$set_state()

teal.slice::FilterState$ui()

Method new()

Initialize a FilterState object. This class has an extra field, private$timezone, which is set to Sys.timezone() by default. However, in case when using this module in teal app, one needs timezone of the app user. App user timezone is taken from session$userData$timezone and is set only if object is initialized in shiny.

Usage

DatetimeFilterState$new(
  x,
  varname,
  varlabel = character(0),
  dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(POSIXct or POSIXlt)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

dataname

(character(1))
optional name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <dataname>$<varname>

"matrix" varname in the condition call will be returned as <dataname>[, <varname>]

Method format()

Returns a formatted string representing this DatetimeFilterState.

Usage

DatetimeFilterState$format(indent = 0)

Arguments

indent

(numeric(1)) the number of spaces before after each new line character of the formatted string. Default: 0

Returns

character(1) the formatted string

Method is_any_filtered()

Answers the question of whether the current settings and values selected actually filters out any values.

Usage

DatetimeFilterState$is_any_filtered()

Returns

logical scalar

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> >= as.POSIXct(<min>) & <varname> <= <max>) with optional is.na(<varname>).

Usage

DatetimeFilterState$get_call()

Method set_selected()

Sets the selected time frame of this DatetimeFilterState.

Usage

DatetimeFilterState$set_selected(value)

Arguments

value

(POSIX(2)) the lower and the upper bound of the selected time frame. Must not contain NA values.

Returns

invisibly NULL.

Examples

date <- as.POSIXct(1, origin = "01/01/1970")
filter <- teal.slice:::DatetimeFilterState$new(
  c(date, date + 1, date + 2, date + 3),
  varname = "name"
)
filter$set_selected(c(date + 1, date + 2))

Method clone()

The objects of this class are cloneable with this method.

Usage

DatetimeFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal.slice:::DatetimeFilterState$new(
  c(Sys.time() + seq(0, by = 3600, length.out = 10), NA),
  varname = "x",
  dataname = "data",
  extract_type = character(0)
)

isolate(filter_state$get_call())
#> x >= as.POSIXct("2023-05-22 18:07:20", tz = "Etc/UTC") & x < 
#>     as.POSIXct("2023-05-23 03:07:21", tz = "Etc/UTC")
isolate(filter_state$set_selected(c(Sys.time() + 3L, Sys.time() + 8L)))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$get_call())
#> is.na(x) | x >= as.POSIXct("2023-05-22 18:07:23", tz = "Etc/UTC") & 
#>     x < as.POSIXct("2023-05-22 18:07:29", tz = "Etc/UTC")

if (FALSE) {
# working filter in an app
library(shiny)

datetimes <- as.POSIXct(c("2012-01-01 12:00:00", "2020-01-01 12:00:00"))
data_datetime <- c(seq(from = datetimes[1], to = datetimes[2], length.out = 100), NA)
filter_state_datetime <- DatetimeFilterState$new(
  x = data_datetime,
  varname = "variable",
  varlabel = "label"
)
filter_state_datetime$set_state(list(selected = data_datetime[c(47, 98)], keep_na = TRUE))

ui <- fluidPage(
  column(4, div(
    h4("DatetimeFilterState"),
    isolate(filter_state_datetime$ui("fs"))
  )),
  column(4, div(
    id = "outputs", # div id is needed for toggling the element
    h4("Condition (i.e. call)"), # display the subsetting call generated by this FilterState
    textOutput("condition_datetime"), br(),
    h4("Unformatted state"), # display raw filter state
    textOutput("unformatted_datetime"), br(),
    h4("Formatted state"), # display human readable filter state
    textOutput("formatted_datetime"), br()
  )),
  column(4, div(
    h4("Programmatic filter control"),
    actionButton("button1_datetime", "set drop NA", width = "100%"), br(),
    actionButton("button2_datetime", "set keep NA", width = "100%"), br(),
    actionButton("button3_datetime", "set a range", width = "100%"), br(),
    actionButton("button4_datetime", "set full range", width = "100%"), br(),
    actionButton("button0_datetime", "set initial state", width = "100%"), br()
  ))
)

server <- function(input, output, session) {
  filter_state_datetime$server("fs")
  output$condition_datetime <- renderPrint(filter_state_datetime$get_call())
  output$formatted_datetime <- renderText(filter_state_datetime$format())
  output$unformatted_datetime <- renderPrint(filter_state_datetime$get_state())
  # modify filter state programmatically
  observeEvent(input$button1_datetime, filter_state_datetime$set_keep_na(FALSE))
  observeEvent(input$button2_datetime, filter_state_datetime$set_keep_na(TRUE))
  observeEvent(
    input$button3_datetime,
    filter_state_datetime$set_selected(data_datetime[c(34, 56)])
  )
  observeEvent(input$button4_datetime, filter_state_datetime$set_selected(datetimes))
  observeEvent(
    input$button0_datetime,
    filter_state_datetime$set_state(list(selected = data_datetime[c(47, 98)], keep_na = TRUE))
  )
}

if (interactive()) {
  shinyApp(ui, server)
}
}

## ------------------------------------------------
## Method `DatetimeFilterState$set_selected`
## ------------------------------------------------

date <- as.POSIXct(1, origin = "01/01/1970")
filter <- teal.slice:::DatetimeFilterState$new(
  c(date, date + 1, date + 2, date + 3),
  varname = "name"
)
filter$set_selected(c(date + 1, date + 2))