DefaultFilteredDataset R6 class

Stores any object as inert entity. Filtering is not supported.

Super class

teal.slice::FilteredDataset -> DefaultFilteredDataset

Methods

Public methods

DefaultFilteredDataset$new()

DefaultFilteredDataset$format()

DefaultFilteredDataset$get_call()

DefaultFilteredDataset$get_filter_state()

DefaultFilteredDataset$set_filter_state()

DefaultFilteredDataset$clear_filter_states()

DefaultFilteredDataset$get_filter_overview()

DefaultFilteredDataset$ui_active()

DefaultFilteredDataset$ui_add()

DefaultFilteredDataset$clone()

Inherited methods

teal.slice::FilteredDataset$destroy()

teal.slice::FilteredDataset$get_dataname()

teal.slice::FilteredDataset$get_dataset()

teal.slice::FilteredDataset$get_dataset_label()

teal.slice::FilteredDataset$get_keys()

teal.slice::FilteredDataset$print()

teal.slice::FilteredDataset$srv_active()

teal.slice::FilteredDataset$srv_add()

Method new()

Initializes this DefaultFilteredDataset object.

Usage

DefaultFilteredDataset$new(dataset, dataname, label = character(0))

Arguments

dataset

any type of object; will not be filtered.

dataname

(character(1)) syntactically valid name given to the dataset.

label

(character(1)) label to describe the dataset.

Returns

Object of class DefaultFilteredDataset, invisibly.

Method format()

Returns a formatted string representing this DefaultFilteredDataset object.

Usage

DefaultFilteredDataset$format(show_all, trim_lines = FALSE)

Arguments

show_all

(logical(1)) for method consistency, ignored.

trim_lines

(logical(1)) flag specifying whether to trim lines if class names are too long.

Returns

The formatted string.

Method get_call()

Usage

DefaultFilteredDataset$get_call(sid)

Arguments

sid

(character(1)) for method consistency, ignored.

Returns

NULL, invisibly.

Method get_filter_state()

Usage

DefaultFilteredDataset$get_filter_state()

Returns

NULL, invisibly.

Method set_filter_state()

Usage

DefaultFilteredDataset$set_filter_state(state)

Arguments

state

(teal_slices) for method consistency, ignored.

Returns

NULL, invisibly.

Method clear_filter_states()

Usage

DefaultFilteredDataset$clear_filter_states(force)

Arguments

force

(logical(1)) for method consistency, ignored.

Returns

NULL, invisibly.

Method get_filter_overview()

Creates row for filter overview in the form of
dataname - unsupported data class

Usage

DefaultFilteredDataset$get_filter_overview()

Returns

A data.frame.

Method ui_active()

Overwrites parent method.

Usage

DefaultFilteredDataset$ui_active(id, allow_add)

Arguments

id

(character(1)) shiny module instance id.

allow_add

(ignored)

Details

Blank UI module that would list active filter states for this dataset.

Returns

An empty div.

Method ui_add()

Overwrites parent method.

Usage

DefaultFilteredDataset$ui_add(id)

Arguments

id

(character(1)) shiny module instance id.

Details

Blank UI module that would list active filter states for this dataset.

Returns

An empty div.

Method clone()

The objects of this class are cloneable with this method.

Usage

DefaultFilteredDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# use non-exported function from teal.slice
DefaultFilteredDataset <- getFromNamespace("DefaultFilteredDataset", "teal.slice")

library(shiny)

ds <- DefaultFilteredDataset$new(letters, "letters")
isolate(ds$get_filter_state())
isolate(ds$get_call())