FilterState object for empty variable

FilterState subclass representing an empty variable

Super class

teal.slice::FilterState -> EmptyFilterState

Methods

Public methods

EmptyFilterState$new()

EmptyFilterState$get_call()

EmptyFilterState$clone()

Inherited methods

teal.slice::FilterState$destroy_observers()

teal.slice::FilterState$format()

teal.slice::FilterState$get_state()

teal.slice::FilterState$print()

teal.slice::FilterState$server()

teal.slice::FilterState$set_state()

teal.slice::FilterState$ui()

Method new()

Initialize EmptyFilterState object.

Usage

EmptyFilterState$new(
  x,
  x_reactive = reactive(NULL),
  extract_type = character(0),
  slice
)

Arguments

x

(vector)
values of the variable used in filter

x_reactive

(reactive)
returning vector of the same type as x. Is used to update counts following the change in values of the filtered dataset. If it is set to reactive(NULL) then counts based on filtered dataset are not shown.

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <dataname>$<varname>

"matrix" varname in the condition call will be returned as <dataname>[, <varname>]

slice

(teal_slice)
object created using teal_slice(). teal_slice is stored in the class and set_state directly manipulates values within teal_slice. get_state returns teal_slice object which can be reused in other places. Beware, that teal_slice is a reactiveValues which means that changes in particular object are automatically reflected in all places which refer to the same teal_slice.

...

additional arguments to be saved as a list in private$extras field

Method get_call()

Returns reproducible condition call for current selection relevant for selected variable type. Uses internal reactive values, hence must be called in reactive or isolated context.

Usage

EmptyFilterState$get_call(dataname)

Arguments

dataname

name of data set; defaults to private$get_dataname()

Returns

logical(1)

Method clone()

The objects of this class are cloneable with this method.

Usage

EmptyFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal.slice:::EmptyFilterState$new(
  x = NA,
  slice = teal_slice(varname = "x", dataname = "data"),
  extract_type = character(0)
)
shiny::isolate(filter_state$get_call())
#> NULL
filter_state$set_state(teal_slice(dataname = "data", varname = "x", keep_na = TRUE))
shiny::isolate(filter_state$get_call())
#> NULL