FilterStateExpr Class

Class to handle filter expression.

Details

This class is responsible for displaying filter card and returning filter expression

Methods

Public methods

FilterStateExpr$new()

FilterStateExpr$format()

FilterStateExpr$print()

FilterStateExpr$get_state()

FilterStateExpr$set_state()

FilterStateExpr$get_call()

FilterStateExpr$destroy_observers()

FilterStateExpr$server()

FilterStateExpr$ui()

FilterStateExpr$clone()

Method new()

Initialize a FilterStateExpr object

Usage

FilterStateExpr$new(slice)

Arguments

slice

(teal_slice_expr)
object created by teal_slice()

Returns

FilterStateExpr

Method format()

Returns a formatted string representing this FilterStateExpr object.

Usage

FilterStateExpr$format(show_all = FALSE, trim_lines = TRUE)

Arguments

show_all

logical(1) passed to format.teal_slice

trim_lines

logical(1) passed to format.teal_slice

Returns

character(1) the formatted string

Method print()

Prints this FilterStateExpr object.

Usage

FilterStateExpr$print(...)

Arguments

...

additional arguments

Method get_state()

Returns filtering state.

Usage

FilterStateExpr$get_state()

Returns

A teal_slice object.

Method set_state()

Sets filtering state.

Usage

FilterStateExpr$set_state(state)

Arguments

state

a teal_slice object

Returns

self invisibly

Method get_call()

Get reproducible call

Usage

FilterStateExpr$get_call(dataname)

Arguments

dataname

(ignored) for a consistency with FilterState

Returns reproducible condition call for current selection relevant for selected variable type. Method is using internal reactive values which makes it reactive and must be executed in reactive or isolated context.

Returns

language

Method destroy_observers()

Destroy observers stored in private$observers.

Usage

FilterStateExpr$destroy_observers()

Returns

NULL invisibly

Method server()

Shiny module server.

Usage

FilterStateExpr$server(id)

Arguments

id

(character(1))
shiny module instance id

Returns

moduleServer function which returns reactive value signaling that remove button has been clicked

Method ui()

Shiny module UI.

Usage

FilterStateExpr$ui(id, parent_id = "cards")

Arguments

id

(character(1))
shiny element (module instance) id; the UI for this class contains simple message stating that it is not supported

parent_id

(character(1)) id of the FilterStates card container

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterStateExpr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal.slice:::FilterStateExpr$new(
  slice = teal_slice(
    dataname = "x",
    id = "FA",
    title = "Adult females",
    expr = "sex == 'F' & age >= 18"
  )
)
filter_state$get_call()
#> sex == "F" & age >= 18

# working filter in an app
library(shiny)
library(shinyjs)

ui <- fluidPage(
  useShinyjs(),
  teal.slice:::include_css_files(pattern = "filter-panel"),
  teal.slice:::include_js_files(pattern = "count-bar-labels"),
  column(4, div(
    h4("ChoicesFilterState"),
    filter_state$ui("fs")
  )),
  column(8, div(
    h4("Condition (i.e. call)"), # display the subsetting call generated by this FilterState
    textOutput("condition_choices"), br(),
    h4("Unformatted state"), # display raw filter state
    textOutput("unformatted_choices"), br(),
    h4("Formatted state"), # display human readable filter state
    textOutput("formatted_choices"), br()
  ))
)

server <- function(input, output, session) {
  filter_state$server("fs")
  output$condition_choices <- renderPrint(filter_state$get_call())
  output$formatted_choices <- renderText(filter_state$format())
  output$unformatted_choices <- renderPrint(filter_state$get_state())
}

if (interactive()) {
  shinyApp(ui, server)
}