FilterStates subclass for MultiAssayExperiments

Handles filter states in a MultiAssayExperiment

Super class

teal.slice::FilterStates -> MAEFilterStates

Methods

Public methods

MAEFilterStates$new()

MAEFilterStates$clone()

Inherited methods

teal.slice::FilterStates$clear_filter_states()

teal.slice::FilterStates$format()

teal.slice::FilterStates$get_call()

teal.slice::FilterStates$get_filter_state()

teal.slice::FilterStates$print()

teal.slice::FilterStates$remove_filter_state()

teal.slice::FilterStates$set_filter_state()

teal.slice::FilterStates$srv_active()

teal.slice::FilterStates$srv_add()

teal.slice::FilterStates$ui_active()

teal.slice::FilterStates$ui_add()

Method new()

Initializes MAEFilterStates object

Initialize MAEFilterStates object

Usage

MAEFilterStates$new(
  data,
  data_reactive = function(sid = "") NULL,
  dataname,
  datalabel = "subjects",
  keys = character(0)
)

Arguments

data

(MultiAssayExperiment)
the R object which MultiAssayExperiment::subsetByColData function is applied on.

data_reactive

(function(sid))
should return a MultiAssayExperiment object or NULL. This object is needed for the FilterState counts being updated on a change in filters. If function returns NULL then filtered counts are not shown. Function has to have sid argument being a character.

dataname

(character(1))
name of the data used in the expression specified to the function argument attached to this FilterStates.

datalabel

(NULL or character(1))
text label value

keys

(character)
key columns names

varlabels

(character)
labels of the variables used in this object

Method clone()

The objects of this class are cloneable with this method.

Usage

MAEFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.