FilterState object for numeric variable

FilterState object for numeric variable

FilterState object for numeric variable

Value

invisibly NULL

Note

Casts the passed object to numeric before validating the input making it possible to pass any object coercible to numeric to this method.

Super class

teal.slice::FilterState -> RangeFilterState

Methods

Public methods

RangeFilterState$new()

RangeFilterState$format()

RangeFilterState$is_any_filtered()

RangeFilterState$get_call()

RangeFilterState$get_keep_inf()

RangeFilterState$get_state()

RangeFilterState$set_keep_inf()

RangeFilterState$set_state()

RangeFilterState$set_selected()

RangeFilterState$clone()

Inherited methods

teal.slice::FilterState$destroy_observers()

teal.slice::FilterState$get_dataname()

teal.slice::FilterState$get_keep_na()

teal.slice::FilterState$get_selected()

teal.slice::FilterState$get_varlabel()

teal.slice::FilterState$get_varname()

teal.slice::FilterState$print()

teal.slice::FilterState$server()

teal.slice::FilterState$set_keep_na()

teal.slice::FilterState$set_na_rm()

teal.slice::FilterState$ui()

Method new()

Initialize a FilterState object

Usage

RangeFilterState$new(
  x,
  varname,
  varlabel = character(0),
  input_dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(numeric)
values of the variable used in filter

varname

(character, name)
name of the variable

varlabel

(character(1))
label of the variable (optional).

input_dataname

(name or call)
name of dataset where x is taken from

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <input_dataname>$<varname>

"matrix" varname in the condition call will be returned as <input_dataname>[, <varname>]

Method format()

Returns a formatted string representing this LogicalFilterState.

Usage

RangeFilterState$format(indent = 0)

Arguments

indent

(numeric(1)) the number of spaces before after each new line character of the formatted string. Default: 0

Returns

character(1) the formatted string

Method is_any_filtered()

Answers the question of whether the current settings and values selected actually filters out any values.

Usage

RangeFilterState$is_any_filtered()

Returns

logical scalar

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> >= <min value> & <varname> <= <max value> with optional is.na(<varname>) and is.finite(<varname>).

Usage

RangeFilterState$get_call()

Returns

(call)

Method get_keep_inf()

Returns current keep_inf selection

Usage

RangeFilterState$get_keep_inf()

Returns

(logical(1))

Method get_state()

Returns the filtering state.

Usage

RangeFilterState$get_state()

Returns

list containing values taken from the reactive fields:

selected (numeric(2)) range of the filter.

keep_na (logical(1)) whether NA should be kept.

keep_inf (logical(1)) whether Inf should be kept.

Method set_keep_inf()

Set if Inf should be kept

Usage

RangeFilterState$set_keep_inf(value)

Arguments

value

(logical(1))
Value(s) which come from the filter selection. Value is set in server modules after selecting check-box-input in the shiny interface. Values are set to private$keep_inf which is reactive.

Method set_state()

Set state

Usage

RangeFilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

keep_inf (logical) defines whether to keep or remove Inf values

Method set_selected()

Sets the selected values of this RangeFilterState.

Usage

RangeFilterState$set_selected(value)

Arguments

value

(numeric(2)) the two-elements array of the lower and upper bound of the selected range. Must not contain NA values.

Examples

filter <- teal.slice:::RangeFilterState$new(c(1, 2, 3, 4), varname = "name")
filter$set_selected(c(2, 3))

Method clone()

The objects of this class are cloneable with this method.

Usage

RangeFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal.slice:::RangeFilterState$new(
  c(NA, Inf, seq(1:10)),
  varname = "x",
  input_dataname = as.name("data"),
  extract_type = character(0)
)
isolate(filter_state$get_call())
#> x >= 1 & x <= 10
isolate(filter_state$set_selected(c(3L, 8L)))
isolate(filter_state$set_keep_na(TRUE))
isolate(filter_state$set_keep_inf(TRUE))
isolate(filter_state$get_call())
#> is.na(x) | (is.infinite(x) | x >= 3 & x <= 8)

## ------------------------------------------------
## Method `RangeFilterState$set_selected`
## ------------------------------------------------

filter <- teal.slice:::RangeFilterState$new(c(1, 2, 3, 4), varname = "name")
filter$set_selected(c(2, 3))