FilterStates subclass for SummarizedExperiments

Handles filter states in a SummaryExperiment.

Super class

teal.slice::FilterStates -> SEFilterStates

Methods

Public methods

SEFilterStates$new()

SEFilterStates$set_filter_state()

SEFilterStates$ui_add()

SEFilterStates$srv_add()

SEFilterStates$clone()

Inherited methods

teal.slice::FilterStates$clear_filter_states()

teal.slice::FilterStates$destroy()

teal.slice::FilterStates$format()

teal.slice::FilterStates$get_call()

teal.slice::FilterStates$get_filter_state()

teal.slice::FilterStates$print()

teal.slice::FilterStates$remove_filter_state()

teal.slice::FilterStates$srv_active()

teal.slice::FilterStates$ui_active()

Method new()

Initialize SEFilterStates object.

Usage

SEFilterStates$new(
  data,
  data_reactive = function(sid = "") NULL,
  dataname,
  datalabel = NULL
)

Arguments

data

(SummarizedExperiment) the R object which subset function is applied on.

data_reactive

(function(sid)) should return a SummarizedExperiment object or NULL. This object is needed for the FilterState counts being updated on a change in filters. If function returns NULL then filtered counts are not shown. Function has to have sid argument being a character.

dataname

(character(1)) name of the data used in the expression specified to the function argument attached to this FilterStates.

datalabel

(character(1)) optional text label. Should be the name of experiment.

Method set_filter_state()

Set filter state.

Usage

SEFilterStates$set_filter_state(state)

Arguments

state

(teal_slices) teal_slice objects should contain the field arg %in% c("subset", "select")

Returns

NULL, invisibly.

Method ui_add()

shiny UI module to add filter variable.

Usage

SEFilterStates$ui_add(id)

Arguments

id

(character(1)) shiny module instance id.

Returns

shiny.tag

Method srv_add()

shiny server module to add filter variable.

Module controls available choices to select as a filter variable. Selected filter variable is being removed from available choices. Removed filter variable gets back to available choices. This module unlike other FilterStates classes manages two sets of filter variables - one for colData and another for rowData.

Usage

SEFilterStates$srv_add(id)

Arguments

id

(character(1)) shiny module instance id.

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

SEFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.