Choices condition call

Compose choices condition call from inputs.

Usage

call_condition_choice(varname, choices)

Arguments

varname

(name, call or character(1))
name of the variable

choices

(vector)
varname values to match using the == (single value) or %in% (vector) condition. choices can be vector of any type but for some output might be converted:

factor call is composed on choices converted to character

Date call is composed on choices converted to character using format(choices)

POSIXct, POSIXlt Call is composed on choices converted to character using format(choices). One has to be careful here as formatted date-time variable might loose some precision (see format argument in format.POSIXlt) and output call could be insufficient for exact comparison. In this case one should specify varname = trunc(<varname>) and possibly convert choices to character)

Value

a call

Examples

teal.slice:::call_condition_choice("SEX", choices = c(1, 2))
#> SEX %in% c(1, 2)
teal.slice:::call_condition_choice(as.name("SEX"), choices = "F")
#> SEX == "F"
teal.slice:::call_condition_choice("SEX", choices = c("F", "M"))
#> SEX %in% c("F", "M")
teal.slice:::call_condition_choice("SEX", choices = factor(c("F", "M")))
#> SEX %in% c("F", "M")
teal.slice:::call_condition_choice("x$SEX", choices = Sys.Date())
#> x$SEX == "2022-05-17"
teal.slice:::call_condition_choice("trunc(x$SEX)", choices = Sys.time())
#> trunc(x$SEX) == "2022-05-17 12:34:48"