FilterStates subclass for data frames

Handles filter states in a data.frame

Super class

teal.slice::FilterStates -> DFFilterStates

Methods

Public methods

DFFilterStates$new()

DFFilterStates$format()

DFFilterStates$get_fun()

DFFilterStates$server()

DFFilterStates$get_filter_state()

DFFilterStates$set_filter_state()

DFFilterStates$remove_filter_state()

DFFilterStates$ui_add_filter_state()

DFFilterStates$srv_add_filter_state()

DFFilterStates$clone()

Inherited methods

teal.slice::FilterStates$get_call()

teal.slice::FilterStates$get_datalabel()

teal.slice::FilterStates$get_filter_count()

teal.slice::FilterStates$print()

teal.slice::FilterStates$state_list_empty()

teal.slice::FilterStates$state_list_get()

teal.slice::FilterStates$state_list_push()

teal.slice::FilterStates$state_list_remove()

teal.slice::FilterStates$ui()

Method new()

Initializes DFFilterStates object.

Initializes DFFilterStates object by setting dataname and initializing state_list (shiny::reactiveVal). This class contains a single state_list with no specified name, which means that when calling the subset function associated with this class (dplyr::filter), a list of conditions is passed to unnamed arguments (...).

Usage

DFFilterStates$new(dataname, datalabel, varlabels, keys)

Arguments

dataname

(character(1))
name of the data used in the subset expression specified to the function argument attached to this FilterStates

datalabel

(character(0) or character(1))
text label value

varlabels

(character)
labels of the variables used in this object

keys

(character)
key columns names

Method format()

Returns a formatted string representing this FilterStates object.

Usage

DFFilterStates$format(indent = 0)

Arguments

indent

(numeric(1)) the number of spaces prepended to each line of the output

Returns

character(1) the formatted string

Method get_fun()

Gets the name of the function used to filter the data in this FilterStates.

Get name of function used to create the subset expression. For DFFilterStates this is dplyr::filter.

Usage

DFFilterStates$get_fun()

Returns

character(1)

Method server()

Shiny server module.

Usage

DFFilterStates$server(id)

Arguments

id

(character(1))
shiny module instance id

Returns

moduleServer function which returns NULL

Method get_filter_state()

Gets the reactive values from the active FilterState objects.

Get active filter state from the FilterState objects kept in state_list. The output list is a compatible input to self$set_filter_state.

Usage

DFFilterStates$get_filter_state()

Returns

list with named elements corresponding to FilterState in the state_list.

Method set_filter_state()

Set filter state.

Usage

DFFilterStates$set_filter_state(
  data,
  state,
  vars_include = get_supported_filter_varnames(data = data),
  ...
)

Arguments

data

(data.frame)
data object for which to define a subset

state

(named list)
should contain values of initial selections in the FilterState; list names must correspond to column names in data

vars_include

(character(n))
optional, vector of column names to be included

...

ignored

Returns

NULL

Examples

dffs <- teal.slice:::DFFilterStates$new(
  dataname = "iris",
  datalabel = character(0),
  varlabels = character(0),
  keys = character(0)
)
fs <- list(
  Sepal.Length = list(selected = c(5.1, 6.4), keep_na = TRUE, keep_inf = TRUE),
  Species = list(selected = c("setosa", "versicolor"), keep_na = FALSE)
)
shiny::isolate(dffs$set_filter_state(state = fs, data = iris))
shiny::isolate(dffs$get_filter_state())

Method remove_filter_state()

Remove a FilterState from the state_list.

Usage

DFFilterStates$remove_filter_state(state_id)

Arguments

state_id

(character(1))
name of state_list element

Returns

NULL

Method ui_add_filter_state()

Shiny UI module to add filter variable.

Usage

DFFilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
shiny element (module instance) id

data

(data.frame)
data object for which to define a subset

Returns

shiny.tag

Method srv_add_filter_state()

Shiny server module to add filter variable.

This module controls available choices to select as a filter variable. Once selected, a variable is removed from available choices. Removing a filter variable adds it back to available choices.

Usage

DFFilterStates$srv_add_filter_state(
  id,
  data,
  vars_include = get_supported_filter_varnames(data = data),
  ...
)

Arguments

id

(character(1))
shiny module instance id

data

(data.frame)
data object for which to define a subset

vars_include

(character(n))
optional, vector of column names to be included

...

ignored

Returns

moduleServer function which returns NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

DFFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# working filters in an app

library(shiny)

# create data frame to filter
data_df <- data.frame(
  NUM1 = 1:100,
  NUM2 = round(runif(100, min = 20, max = 23)),
  CHAR1 = sample(LETTERS[1:6], size = 100, replace = TRUE),
  CHAR2 = sample(c("M", "F"), size = 100, replace = TRUE),
  DATE = seq(as.Date("2020-01-01"), by = 1, length.out = 100),
  DATETIME = as.POSIXct(seq(as.Date("2020-01-01"), by = 1, length.out = 100))
)
data_na <- data.frame(
  NUM1 = NA,
  NUM2 = NA,
  CHAR1 = NA,
  CHAR2 = NA,
  DATE = NA,
  DATETIME = NA
)
data_df <- rbind(data_df, data_na)

# initiate FilterStates object
filter_states_df <- init_filter_states(
  data = data_df,
  dataname = "dataset",
  datalabel = ("label"),
  varlabels = c("long", "short", "long", "short", "long", "long")
)

ui <- fluidPage(
  column(4, div(
    h4("Active filters"),
    filter_states_df$ui("fsdf")
  )),
  column(4, div(
    h4("Manual filter control"),
    filter_states_df$ui_add_filter_state("add_filters", data_df), br(),
    h4("Condition (i.e. call)"), # display the subsetting call generated by this FilterStates
    textOutput("call_df"), br(),
    h4("Formatted state"), # display human readable filter state
    textOutput("formatted_df"), br()
  )),
  column(4, div(
    h4("Programmatic filter control"),
    actionButton("button1_df", "set NUM1 < 30", width = "100%"), br(),
    actionButton("button2_df", "set NUM2 %in% c(20, 21)", width = "100%"), br(),
    actionButton("button3_df", "set CHAR1 %in% c(\"B\", \"C\", \"D\")", width = "100%"), br(),
    actionButton("button4_df", "set CHAR2 == \"F\"", width = "100%"), br(),
    actionButton("button5_df", "set DATE <= 2020-02-02", width = "100%"), br(),
    actionButton("button6_df", "set DATETIME <= 2020-02-02", width = "100%"), br(),
    hr(),
    actionButton("button7_df", "remove NUM1", width = "100%"), br(),
    actionButton("button8_df", "remove NUM2", width = "100%"), br(),
    actionButton("button9_df", "remove CHAR1", width = "100%"), br(),
    actionButton("button10_df", "remove CHAR2", width = "100%"), br(),
    actionButton("button11_df", "remove DATE", width = "100%"), br(),
    actionButton("button12_df", "remove DATETIME", width = "100%"), br(),
    hr(),
    actionButton("button0_df", "clear all filters", width = "100%"), br()
  ))
)

server <- function(input, output, session) {
  filter_states_df$srv_add_filter_state("add_filters", data_df)
  filter_states_df$server("fsdf")

  output$call_df <- renderPrint(filter_states_df$get_call())
  output$formatted_df <- renderText(filter_states_df$format())

  observeEvent(input$button1_df, {
    filter_state <- list(NUM1 = list(selected = c(0, 30)))
    filter_states_df$set_filter_state(data = data_df, state = filter_state)
  })
  observeEvent(input$button2_df, {
    filter_state <- list(NUM2 = list(selected = c(20, 21)))
    filter_states_df$set_filter_state(data = data_df, state = filter_state)
  })
  observeEvent(input$button3_df, {
    filter_state <- list(CHAR1 = list(selected = c("B", "C", "D")))
    filter_states_df$set_filter_state(data = data_df, state = filter_state)
  })
  observeEvent(input$button4_df, {
    filter_state <- list(CHAR2 = list(selected = "F"))
    filter_states_df$set_filter_state(data = data_df, state = filter_state)
  })
  observeEvent(input$button5_df, {
    filter_state <- list(DATE = list(selected = c("2020-01-01", "2020-02-02")))
    filter_states_df$set_filter_state(data = data_df, state = filter_state)
  })
  observeEvent(input$button6_df, {
    filter_state <- list(DATETIME = list(selected = as.POSIXct(c("2020-01-01", "2020-02-02"))))
    filter_states_df$set_filter_state(data = data_df, state = filter_state)
  })
  observeEvent(input$button7_df, filter_states_df$state_list_remove(1, state_id = "NUM1"))
  observeEvent(input$button8_df, filter_states_df$state_list_remove(1, state_id = "NUM2"))
  observeEvent(input$button9_df, filter_states_df$state_list_remove(1, state_id = "CHAR1"))
  observeEvent(input$button10_df, filter_states_df$state_list_remove(1, state_id = "CHAR2"))
  observeEvent(input$button11_df, filter_states_df$state_list_remove(1, state_id = "DATE"))
  observeEvent(input$button12_df, filter_states_df$state_list_remove(1, state_id = "DATETIME"))
  observeEvent(input$button0_df, filter_states_df$state_list_empty())
}

if (interactive()) {
  shinyApp(ui, server)
}
}

## ------------------------------------------------
## Method `DFFilterStates$set_filter_state`
## ------------------------------------------------

dffs <- teal.slice:::DFFilterStates$new(
  dataname = "iris",
  datalabel = character(0),
  varlabels = character(0),
  keys = character(0)
)
fs <- list(
  Sepal.Length = list(selected = c(5.1, 6.4), keep_na = TRUE, keep_inf = TRUE),
  Species = list(selected = c("setosa", "versicolor"), keep_na = FALSE)
)
shiny::isolate(dffs$set_filter_state(state = fs, data = iris))
#> NULL
shiny::isolate(dffs$get_filter_state())
#> $Sepal.Length
#> $Sepal.Length$selected
#> [1] 5.1 6.4
#> 
#> $Sepal.Length$keep_na
#> [1] TRUE
#> 
#> $Sepal.Length$keep_inf
#> [1] TRUE
#> 
#> 
#> $Species
#> $Species$selected
#> [1] "setosa"     "versicolor"
#> 
#> $Species$keep_na
#> [1] FALSE
#> 
#>