FilterState Abstract Class

Abstract class to encapsulate filter states

Details

This class is responsible for managing single filter item within FilteredData class. Filter states depend on the variable type: (logical, integer, numeric, factor, character, Date, POSIXct, POSIXlt) and returns FilterState object with class corresponding to input variable. Class controls single filter entry in module_single_filter_item and returns code relevant to selected values.

factor, character: class = ChoicesFilterState

numeric: class = RangeFilterState

logical: class = LogicalFilterState

Date: class = DateFilterState

POSIXct, POSIXlt: class = DatetimeFilterState

all NA entries: class: FilterState, cannot be filtered

default: FilterState, cannot be filtered
Each variable's filter state is an R6 object which contains choices, selected, varname, dataname, labels, na_count, keep_na and other variable type specific fields (keep_inf, inf_count, timezone). Object contains also shiny module (ui and server) which manages state of the filter through reactive values selected, keep_na, keep_inf which trigger get_call() and every R function call up in reactive chain.

Modifying state

Modifying a FilterState object is possible in three scenarios:

In the interactive session by directly specifying values of selected, keep_na or keep_inf using set_state method (to update all at once), or using set_selected, set_keep_na or set_keep_inf

In a running application by changing appropriate inputs

In a running application by using filter_state_api which directly uses set_state method of the FilterState object.

Methods

Public methods

FilterState$new()

FilterState$destroy_observers()

FilterState$format()

FilterState$get_call()

FilterState$get_dataname()

FilterState$get_keep_na()

FilterState$get_varlabel()

FilterState$get_varname()

FilterState$get_selected()

FilterState$get_state()

FilterState$print()

FilterState$set_keep_na()

FilterState$set_na_rm()

FilterState$set_selected()

FilterState$set_state()

FilterState$server()

FilterState$ui()

FilterState$clone()

Method new()

Initialize a FilterState object

Usage

FilterState$new(
  x,
  varname,
  varlabel = character(0),
  dataname = NULL,
  extract_type = character(0)
)

Arguments

x

(vector)
values of the variable used in filter

varname

(character)
name of the variable

varlabel

(character(1))
label of the variable (optional).

dataname

(character(1))
name of dataset where x is taken from. Must be specified if extract_type argument is not empty.

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <dataname>$<varname>

"matrix" varname in the condition call will be returned as <dataname>[, <varname>]

Returns

self invisibly

Method destroy_observers()

Destroy observers stored in private$observers.

Usage

FilterState$destroy_observers()

Returns

NULL invisibly

Method format()

Returns a formatted string representing this FilterState.

Usage

FilterState$format(indent = 0L, wrap_width = 76L)

Arguments

indent

(numeric(1)) number of spaces before after each new line character of the formatted string; defaults to 0

wrap_width

(numeric(1)) number of characters to wrap lines at in the printed output; allowed range is 30 to 120; defaults to 76

Returns

character(1) the formatted string

Method get_call()

Returns reproducible condition call for current selection relevant for selected variable type. Method is using internal reactive values which makes it reactive and must be executed in reactive or isolated context.

Usage

FilterState$get_call()

Method get_dataname()

Returns dataname or "NULL" if dataname is NULL.

Usage

FilterState$get_dataname()

Returns

character(1)

Method get_keep_na()

Returns current keep_na selection.

Usage

FilterState$get_keep_na()

Returns

logical(1)

Method get_varlabel()

Returns variable label.

Usage

FilterState$get_varlabel()

Returns

character(1)

Method get_varname()

Get variable name.

Usage

FilterState$get_varname()

Returns

character(1)

Method get_selected()

Get selected values from FilterState.

Usage

FilterState$get_selected()

Returns

class of the returned object depends of class of the FilterState

Method get_state()

Returns the filtering state.

Usage

FilterState$get_state()

Returns

list containing values taken from the reactive fields:

selected (atomic) length depends on a FilterState variant.

keep_na (logical(1)) whether NA should be kept.

Method print()

Prints this FilterState object.

Usage

FilterState$print(...)

Arguments

...

additional arguments to this method

Method set_keep_na()

Set whether to keep NAs.

Usage

FilterState$set_keep_na(value)

Arguments

value

logical(1)
value(s) which come from the filter selection. Value is set in server modules after selecting check-box-input in the shiny interface. Values are set to private$keep_na which is reactive.

Returns

NULL invisibly

Method set_na_rm()

Some methods need an additional !is.na(varame) condition to drop missing values. When private$na_rm = TRUE, self$get_call returns condition extended by !is.na.

Usage

FilterState$set_na_rm(value)

Arguments

value

logical(1)
when TRUE, FilterState$get_call appends an expression removing NA values to the filter expression returned by get_call

Returns

NULL invisibly

Method set_selected()

Set selection.

Usage

FilterState$set_selected(value)

Arguments

value

(vector)
value(s) that come from filter selection; values are set in the module server after a selection is made in the app interface; values are stored in private$selected which is reactive; value types have to be the same as private$choices

Returns

NULL invisibly

Method set_state()

Set state.

Usage

FilterState$set_state(state)

Arguments

state

(list)
contains fields relevant for a specific class:

selected defines initial selection

keep_na (logical) defines whether to keep or remove NA values

Returns

NULL invisibly

Method server()

Shiny module server.

Usage

FilterState$server(id)

Arguments

id

(character(1))
shiny module instance id

Returns

moduleServer function which returns reactive value signaling that remove button has been clicked

Method ui()

Shiny module UI.

Usage

FilterState$ui(id)

Arguments

id

(character(1))
shiny element (module instance) id; the UI for this class contains simple message stating that it is not supported

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.