FilterState Abstract Class

Abstract class to encapsulate single filter state

Details

This class is responsible for managing single filter item within FilteredData class. Filter states depend on the variable type: (logical, integer, numeric, factor, character, Date, POSIXct, POSIXlt) and returns FilterState object with class corresponding to input variable. Class controls single filter entry in module_single_filter_item and returns code relevant to selected values.

factor, character: class = ChoicesFilterState

numeric: class = RangeFilterState

logical: class = LogicalFilterState

Date: class = DateFilterState

POSIXct, POSIXlt: class = DatetimeFilterState

all NA entries: class: FilterState, cannot be filtered

default: FilterState, cannot be filtered
Each variable's filter state is an R6 object which contains choices, selected, varname, dataname, labels, na_count, keep_na and other variable type specific fields (keep_inf, inf_count, timezone). Object contains also shiny module (ui and server) which manages state of the filter through reactive values selected, keep_na, keep_inf which trigger get_call() and every R function call up in reactive chain.

Modifying state

Modifying a FilterState object is possible in three scenarios:

In the interactive session by passing an appropriate teal_slice to the set_state method, or using set_selected, set_keep_na or set_keep_inf methods.

In a running application by changing appropriate inputs.

In a running application by using filter_state_api which directly uses set_state method of the InteractiveFilterState object.

Methods

Public methods

FilterState$new()

FilterState$format()

FilterState$print()

FilterState$set_state()

FilterState$get_state()

FilterState$get_call()

FilterState$server()

FilterState$ui()

FilterState$destroy_observers()

FilterState$clone()

Method new()

Initialize a FilterState object

Usage

FilterState$new(
  x,
  x_reactive = reactive(NULL),
  slice,
  extract_type = character(0)
)

Arguments

x

(vector)
values of the variable used in filter

x_reactive

(reactive)
returning vector of the same type as x. Is used to update counts following the change in values of the filtered dataset. If it is set to reactive(NULL) then counts based on filtered dataset are not shown.

slice

(teal_slice)
object created by teal_slice()

extract_type

(character(0), character(1))
specifying whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <dataname>$<varname>

"matrix" varname in the condition call will be returned as <dataname>[, <varname>]

...

additional arguments to be saved as a list in private$extras field

Returns

self invisibly

Method format()

Returns a formatted string representing this FilterState object.

Usage

FilterState$format(show_all = FALSE, trim_lines = TRUE)

Arguments

show_all

logical(1) passed to format.teal_slice

trim_lines

logical(1) passed to format.teal_slice

Returns

character(1) the formatted string

Method print()

Prints this FilterState object.

Usage

FilterState$print(...)

Arguments

...

additional arguments

Method set_state()

Sets filtering state.

fixed state is prevented from changing state

locked state is prevented from removing state

Usage

FilterState$set_state(state)

Arguments

state

a teal_slice object

Returns

self invisibly

Method get_state()

Returns filtering state.

Usage

FilterState$get_state()

Returns

A teal_slice object.

Method get_call()

Returns reproducible condition call for current selection relevant for selected variable type. Method is using internal reactive values which makes it reactive and must be executed in reactive or isolated context.

Usage

FilterState$get_call()

Method server()

Shiny module server.

Usage

FilterState$server(id)

Arguments

id

(character(1))
shiny module instance id

Returns

moduleServer function which returns reactive value signaling that remove button has been clicked

Method ui()

Shiny module UI.

Usage

FilterState$ui(id, parent_id = "cards")

Arguments

id

(character(1))
shiny element (module instance) id; the UI for this class contains simple message stating that it is not supported

parent_id

(character(1)) id of the FilterStates card container

Method destroy_observers()

Destroy observers stored in private$observers.

Usage

FilterState$destroy_observers()

Returns

NULL invisibly

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.