FilterStates R6 class

Abstract class that manages adding and removing FilterState objects and builds a subset expression.

A FilterStates object tracks all subsetting expressions (logical predicates that limit observations) associated with a given dataset and composes them into a single reproducible R expression that will assign a subset of the original data to a new variable. This expression is hereafter referred to as subset expression.

The subset expression is constructed differently for different classes of the underlying data object and FilterStates subclasses. Currently implemented for data.frame, matrix, SummarizedExperiment, and MultiAssayExperiment.

Methods

Public methods

FilterStates$new()

FilterStates$get_datalabel()

FilterStates$format()

FilterStates$get_call()

FilterStates$print()

FilterStates$get_fun()

FilterStates$state_list_initialize()

FilterStates$state_list_get()

FilterStates$state_list_push()

FilterStates$state_list_remove()

FilterStates$state_list_empty()

FilterStates$get_filter_count()

FilterStates$remove_filter_state()

FilterStates$ui()

FilterStates$get_filter_state()

FilterStates$set_filter_state()

FilterStates$ui_add_filter_state()

FilterStates$srv_add_filter_state()

FilterStates$clone()

Method new()

Initializes FilterStates object.

Initializes FilterStates object by setting input_dataname, output_dataname and initializing state_list (list of reactiveVals).

Usage

FilterStates$new(input_dataname, output_dataname, datalabel)

Arguments

input_dataname

(character(1) or name or call)
name of the data used on rhs of the expression specified to the function argument attached to this FilterStates

output_dataname

(character(1) or name or call)
name of the output data on the lhs of the subset expression

datalabel

(character(0) or character(1))
text label value

Returns

self invisibly

Method get_datalabel()

Returns the label of the dataset.

Usage

FilterStates$get_datalabel()

Returns

character(1) the data label

Method format()

Returns a formatted string representing this FilterStates object.

Usage

FilterStates$format(indent)

Arguments

indent

(numeric(1)) the number of spaces prepended to each line of the output

Returns

character(1) the formatted string

Method get_call()

Filter call

Builds subset expression from condition calls stored in FilterState objects selection. The lhs of the expression is private$output_dataname. The rhs is a call to self$get_fun() with private$input_dataname as argument and a list of condition calls from FilterState objects stored in private$state_list. If input_dataname is the same as output_dataname and no filters are applied, NULL is returned to avoid no-op calls such as x <- x.

Usage

FilterStates$get_call()

Returns

call or NULL

Method print()

Prints this FilterStates object.

Usage

FilterStates$print(...)

Arguments

...

additional arguments to this method

Method get_fun()

Gets the name of the function used to filter the data in this FilterStates.

Get name of function used to create the subset expression. Defaults to "subset" but can be overridden by child class method.

Usage

FilterStates$get_fun()

Returns

character(1) the name of the function

Method state_list_initialize()

Populates the private state_list field to store active filter states.

Usage

FilterStates$state_list_initialize(x)

Arguments

x

(list of reactiveVal objects)

Method state_list_get()

Returns a list of FilterState objects stored in this FilterStates.

Usage

FilterStates$state_list_get(state_list_index, state_id = NULL)

Arguments

state_list_index

(character(1), integer(1))
index on the list in private$state_list where filter states are kept

state_id

(character(1))
name of element in a filter state (which is a reactiveVal containing a list)

Returns

list of FilterState objects

Method state_list_push()

Adds a new FilterState object to this FilterStates.
Raises error if the length of x does not match the length of state_id.

Usage

FilterStates$state_list_push(x, state_list_index, state_id)

Arguments

x

(FilterState)
object to be added to filter state list

state_list_index

(character(1), integer(1))
index on the list in private$state_list where filter states are kept

state_id

(character(1))
name of element in a filter state (which is a reactiveVal containing a list)

Returns

NULL

Method state_list_remove()

Removes a single filter state with all associated shiny elements:

specified FilterState from private$state_list

UI card created for this filter

observers tracking the selection and remove button

Usage

FilterStates$state_list_remove(state_list_index, state_id)

Arguments

state_list_index

(character(1), integer(1))
index on the list in private$state_list where filter states are kept

state_id

(character(1))
name of element in a filter state (which is a reactiveVal containing a list)

Returns

NULL

Method state_list_empty()

Remove all FilterState objects from this FilterStates object.

Usage

FilterStates$state_list_empty()

Returns

NULL

Method get_filter_count()

Gets the number of active FilterState objects in this FilterStates object.

Usage

FilterStates$get_filter_count()

Returns

integer(1)

Method remove_filter_state()

Remove a single FilterState from state_list.

Usage

FilterStates$remove_filter_state(state_id)

Arguments

state_id

(character)
name of variable for which to remove FilterState

Returns

NULL

Method ui()

Shiny module UI

Shiny UI element that stores FilterState UI elements. Populated with elements created with renderUI in the module server.

Usage

FilterStates$ui(id)

Arguments

id

(character(1))
shiny element (module instance) id

Returns

shiny.tag

Method get_filter_state()

Gets reactive values from active FilterState objects.

Get active filter state from FilterState objects stored in state_list(s). The output is a list compatible with input to self$set_filter_state.

Usage

FilterStates$get_filter_state()

Returns

list containing list per FilterState in the state_list

Method set_filter_state()

Sets active FilterState objects.

Usage

FilterStates$set_filter_state(data, state, filtered_dataset)

Arguments

data

(data.frame)
data object for which to define a subset

state

(named list)
should contain values of initial selections in the FilterState; list names must correspond to column names in data

filtered_dataset

data object for which to define a subset(?)

Method ui_add_filter_state()

Shiny module UI that adds a filter variable.

Usage

FilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
shiny element (module instance) id

data

(data.frame, MultiAssayExperiment, SummarizedExperiment, matrix)
data object for which to define a subset

Returns

shiny.tag

Method srv_add_filter_state()

Shiny module server that adds a filter variable.

Usage

FilterStates$srv_add_filter_state(id, data, ...)

Arguments

id

(character(1))
shiny module instance id

data

(data.frame, MultiAssayExperiment, SummarizedExperiment, matrix)
data object for which to define a subset

...

ignored

Returns

moduleServer function which returns NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(shiny)
filter_states <- teal.slice:::DFFilterStates$new(
  input_dataname = "data",
  output_dataname = "data_filtered",
  varlabels = c(x = "x variable", SEX = "Sex"),
  datalabel = character(0),
  keys = character(0)
)
filter_state <- teal.slice:::RangeFilterState$new(
  c(NA, Inf, seq(1:10)),
  varname = "x",
  varlabel = "x variable",
  input_dataname = as.name("data"),
  extract_type = "list"
)
isolate(filter_state$set_selected(c(3L, 8L)))

isolate(
  filter_states$state_list_push(
    x = filter_state,
    state_list_index = 1L,
    state_id = "x"
  ))
isolate(filter_states$get_call())
#> data_filtered <- dplyr::filter(data, data$x >= 3 & data$x <= 
#>     8)