MAEFilterStates

Specialization of FilterStates for MultiAssayExperiment.

Super class

teal.slice::FilterStates -> MAEFilterStates

Methods

Public methods

MAEFilterStates$new()

MAEFilterStates$format()

MAEFilterStates$get_fun()

MAEFilterStates$server()

MAEFilterStates$get_filter_state()

MAEFilterStates$set_filter_state()

MAEFilterStates$remove_filter_state()

MAEFilterStates$ui_add_filter_state()

MAEFilterStates$srv_add_filter_state()

MAEFilterStates$clone()

Inherited methods

teal.slice::FilterStates$get_call()

teal.slice::FilterStates$get_datalabel()

teal.slice::FilterStates$print()

teal.slice::FilterStates$queue_empty()

teal.slice::FilterStates$queue_get()

teal.slice::FilterStates$queue_initialize()

teal.slice::FilterStates$queue_push()

teal.slice::FilterStates$queue_remove()

teal.slice::FilterStates$ui()

Method new()

Initializes MAEFilterStates object

Initialize MAEFilterStates object

Usage

MAEFilterStates$new(
  input_dataname,
  output_dataname,
  datalabel,
  varlabels,
  keys
)

Arguments

input_dataname

(character(1) or name or call)
name of the data used on lhs of the expression specified to the function argument attached to this FilterStates.

output_dataname

(character(1) or name or call)
name of the output data on the lhs of the assignment expression.

datalabel

(character(0) or character(1))
text label value.

varlabels

(character)
labels of the variables used in this object

keys

(character)
key columns names

Method format()

Returns the formatted string representing this MAEFilterStates object.

Usage

MAEFilterStates$format(indent = 0)

Arguments

indent

(numeric(1)) the number of spaces before each line of the representation

Returns

character(1) the formatted string

Method get_fun()

Returns function name used to create filter call. For MAEFilterStates MultiAssayExperiment::subsetByColData is used.

Usage

MAEFilterStates$get_fun()

Returns

character(1)

Method server()

Server module

Usage

MAEFilterStates$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method get_filter_state()

Returns active FilterState objects.

Gets all active filters from this dataset in form of the nested list. The output list can be used as input to self$set_filter_state.

Usage

MAEFilterStates$get_filter_state()

Returns

list with elements number equal number of FilterStates.

Method set_filter_state()

Set filter state

Usage

MAEFilterStates$set_filter_state(data, state, ...)

Arguments

data

(MultiAssayExperiment)
data which are supposed to be filtered.

state

(named list)
should contain values which are initial selection in the FilterState. Names of the list element should correspond to the name of the column in colData(data).

...

ignored.

Returns

NULL

Method remove_filter_state()

Removes a variable from the ReactiveQueue and its corresponding UI element.

Usage

MAEFilterStates$remove_filter_state(element_id)

Arguments

element_id

(character(1))
name of ReactiveQueue element.

Returns

NULL

Method ui_add_filter_state()

Shiny UI module to add filter variable

Usage

MAEFilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
id of shiny module

data

(MultiAssayExperiment)
object containing colData which columns are used to be used to choose filter variables.

Returns

shiny.tag

Method srv_add_filter_state()

Shiny server module to add filter variable.

Module controls available choices to select as a filter variable. Selected filter variable is being removed from available choices. Removed filter variable gets back to available choices.

Usage

MAEFilterStates$srv_add_filter_state(id, data, ...)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

data

(MultiAssayExperiment)
object containing colData which columns are used to choose filter variables in teal.widgets::optionalSelectInput().

...

ignored

Returns

moduleServer function which returns NULL description Get label of specific variable. In case when variable label is missing name of the variable is returned. parameter variable (character)
name of the variable for which label should be returned return character

Method clone()

The objects of this class are cloneable with this method.

Usage

MAEFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.