MAEFilteredDataset R6 class

MAEFilteredDataset R6 class

MAEFilteredDataset R6 class

Super class

teal.slice::FilteredDataset -> MAEFilteredDataset

Methods

Public methods

MAEFilteredDataset$new()

MAEFilteredDataset$set_filter_state()

MAEFilteredDataset$remove_filter_state()

MAEFilteredDataset$ui_add()

MAEFilteredDataset$get_filter_overview()

MAEFilteredDataset$clone()

Inherited methods

teal.slice::FilteredDataset$clear_filter_states()

teal.slice::FilteredDataset$format()

teal.slice::FilteredDataset$get_call()

teal.slice::FilteredDataset$get_dataname()

teal.slice::FilteredDataset$get_dataset()

teal.slice::FilteredDataset$get_dataset_label()

teal.slice::FilteredDataset$get_filter_count()

teal.slice::FilteredDataset$get_filter_state()

teal.slice::FilteredDataset$get_keys()

teal.slice::FilteredDataset$get_metadata()

teal.slice::FilteredDataset$print()

teal.slice::FilteredDataset$srv_active()

teal.slice::FilteredDataset$srv_add()

teal.slice::FilteredDataset$ui_active()

Method new()

Initialize MAEFilteredDataset object

Usage

MAEFilteredDataset$new(
  dataset,
  dataname,
  keys = character(0),
  label = character(0),
  metadata = NULL
)

Arguments

dataset

(MulitiAssayExperiment)
a single MultiAssayExperiment for which to define a subset

dataname

(character)
a given name for the dataset it may not contain spaces

keys

optional, (character)
vector with primary keys

label

(character)
label to describe the dataset

metadata

(named list or NULL)
field containing metadata about the dataset; each element of the list must be atomic and length one

Method set_filter_state()

Set filter state

Usage

MAEFilteredDataset$set_filter_state(state)

Arguments

state

(named list)
names of the list should correspond to the names of the initialized FilterStates kept in private$filter_states. For this object they are "subjects" and names of the experiments. Values of initial state should be relevant to the referred column.

Returns

NULL invisibly

Examples

utils::data(miniACC, package = "MultiAssayExperiment")
dataset <- teal.slice:::MAEFilteredDataset$new(miniACC, "MAE")
fs <- teal_slices(
  teal_slice(
    dataname = "MAE", varname = "years_to_birth", selected = c(30, 50), keep_na = TRUE
  ),
  teal_slice(
    dataname = "MAE", varname = "vital_status", selected = "1", keep_na = FALSE
  ),
  teal_slice(
    dataname = "MAE", varname = "gender", selected = "female", keep_na = TRUE
  ),
  teal_slice(
    dataname = "MAE", varname = "ARRAY_TYPE", selected = "", keep_na = TRUE
  )
)
dataset$set_filter_state(state = fs)
shiny::isolate(dataset$get_filter_state())

Method remove_filter_state()

Remove one or more FilterState of a MAEFilteredDataset

Usage

MAEFilteredDataset$remove_filter_state(state)

Arguments

state

(teal_slices)
specifying FilterState objects to remove; teal_slices may contain only dataname and varname, other elements are ignored

Returns

NULL invisibly

Method ui_add()

UI module to add filter variable for this dataset

UI module to add filter variable for this dataset

Usage

MAEFilteredDataset$ui_add(id)

Arguments

id

(character(1))
identifier of the element - preferably containing dataset name

Returns

function - shiny UI module

Method get_filter_overview()

Get filter overview rows of a dataset

Usage

MAEFilteredDataset$get_filter_overview()

Returns

(matrix) matrix of observations and subjects

Method clone()

The objects of this class are cloneable with this method.

Usage

MAEFilteredDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `MAEFilteredDataset$set_filter_state`
## ------------------------------------------------

utils::data(miniACC, package = "MultiAssayExperiment")
dataset <- teal.slice:::MAEFilteredDataset$new(miniACC, "MAE")
fs <- teal_slices(
  teal_slice(
    dataname = "MAE", varname = "years_to_birth", selected = c(30, 50), keep_na = TRUE
  ),
  teal_slice(
    dataname = "MAE", varname = "vital_status", selected = "1", keep_na = FALSE
  ),
  teal_slice(
    dataname = "MAE", varname = "gender", selected = "female", keep_na = TRUE
  ),
  teal_slice(
    dataname = "MAE", varname = "ARRAY_TYPE", selected = "", keep_na = TRUE
  )
)
dataset$set_filter_state(state = fs)
#> [WARN] 2023-08-01 07:10:24.0169 pid:642 token:[] teal.slice filters for columns: ARRAY_TYPE excluded from MAE
shiny::isolate(dataset$get_filter_state())
#> {
#>   "slices": [
#>     {
#>       "dataname"       : "MAE",
#>       "varname"        : "years_to_birth",
#>       "id"             : "MAE years_to_birth",
#>       "choices"        : [14, 83],
#>       "selected"       : [30, 50],
#>       "keep_na"        : true,
#>       "fixed"          : false,
#>       "anchored"       : false,
#>       "multiple"       : true
#>     },
#>     {
#>       "dataname"       : "MAE",
#>       "varname"        : "vital_status",
#>       "id"             : "MAE vital_status",
#>       "choices"        : ["0", "1"],
#>       "selected"       : ["1"],
#>       "keep_na"        : false,
#>       "fixed"          : false,
#>       "anchored"       : false,
#>       "multiple"       : true
#>     },
#>     {
#>       "dataname"       : "MAE",
#>       "varname"        : "gender",
#>       "id"             : "MAE gender",
#>       "choices"        : ["female", "male"],
#>       "selected"       : ["female"],
#>       "keep_na"        : true,
#>       "fixed"          : false,
#>       "anchored"       : false,
#>       "multiple"       : true
#>     }
#>   ],
#>   "attributes": {
#>     "include_varnames" : {
#>       "MAE"            : ["patientID", "years_to_birth", ...
#>     },
#>     "count_type"       : "none",
#>     "allow_add"        : true
#>   }
#> }