FilterStates subclass for matrices

Handles filter states in a matrix.

Super class

teal.slice::FilterStates -> MatrixFilterStates

Methods

Public methods

MatrixFilterStates$new()

MatrixFilterStates$clone()

Inherited methods

teal.slice::FilterStates$clear_filter_states()

teal.slice::FilterStates$finalize()

teal.slice::FilterStates$format()

teal.slice::FilterStates$get_call()

teal.slice::FilterStates$get_filter_state()

teal.slice::FilterStates$print()

teal.slice::FilterStates$remove_filter_state()

teal.slice::FilterStates$set_filter_state()

teal.slice::FilterStates$srv_active()

teal.slice::FilterStates$srv_add()

teal.slice::FilterStates$ui_active()

teal.slice::FilterStates$ui_add()

Method new()

Initialize MatrixFilterStates object.

Usage

MatrixFilterStates$new(
  data,
  data_reactive = function(sid = "") NULL,
  dataname,
  datalabel = NULL
)

Arguments

data

(matrix) the R object which subset function is applied on.

data_reactive

(function(sid)) should return a matrix object or NULL. This object is needed for the FilterState counts being updated on a change in filters. If function returns NULL then filtered counts are not shown. Function has to have sid argument being a character.

dataname

(character(1)) name of the data used in the subset expression. Passed to the function argument attached to this FilterStates.

datalabel

(character(1)) optional text label. Should be a name of experiment.

Method clone()

The objects of this class are cloneable with this method.

Usage

MatrixFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.