FilterStates subclass for matrices

Handles filter states in a matrix

Super class

teal.slice::FilterStates -> MatrixFilterStates

Methods

Public methods

MatrixFilterStates$new()

MatrixFilterStates$format()

MatrixFilterStates$server()

MatrixFilterStates$get_filter_state()

MatrixFilterStates$set_filter_state()

MatrixFilterStates$remove_filter_state()

MatrixFilterStates$ui_add_filter_state()

MatrixFilterStates$srv_add_filter_state()

MatrixFilterStates$clone()

Inherited methods

teal.slice::FilterStates$get_call()

teal.slice::FilterStates$get_datalabel()

teal.slice::FilterStates$get_filter_count()

teal.slice::FilterStates$get_fun()

teal.slice::FilterStates$print()

teal.slice::FilterStates$state_list_empty()

teal.slice::FilterStates$state_list_get()

teal.slice::FilterStates$state_list_push()

teal.slice::FilterStates$state_list_remove()

teal.slice::FilterStates$ui()

Method new()

Initialize MatrixFilterStates object

Initialize MatrixFilterStates object

Usage

MatrixFilterStates$new(dataname, datalabel)

Arguments

dataname

(character(1))
name of the data used in the expression specified to the function argument attached to this FilterStates.

datalabel

(character(0) or character(1))
text label value.

Method format()

Returns the formatted string representing this MatrixFilterStates object.

Usage

MatrixFilterStates$format(indent = 0)

Arguments

indent

(numeric(1)) the number of spaces before each line of the representation

Returns

character(1) the formatted string

Method server()

Server module

Usage

MatrixFilterStates$server(id)

Arguments

id

(character(1))
an ID string that corresponds with the ID used to call the module's UI function.

Returns

moduleServer function which returns NULL

Method get_filter_state()

Returns active FilterState objects.

Gets all active filters from this dataset in form of the nested list. The output list can be used as input to self$set_filter_state.

Usage

MatrixFilterStates$get_filter_state()

Returns

list containing list with selected values for each FilterState.

Method set_filter_state()

Sets a filter state

Usage

MatrixFilterStates$set_filter_state(data, state, ...)

Arguments

data

(matrix)
data which are supposed to be filtered.

state

(named list)
should contain values which are initial selection in the FilterState. Names of the list element should correspond to the name of the column in data.

...

ignored.

Returns

NULL

Method remove_filter_state()

Remove a variable from the state_list and its corresponding UI element.

Usage

MatrixFilterStates$remove_filter_state(state_id)

Arguments

state_id

(character(1))
name of state_list element.

Returns

NULL

Method ui_add_filter_state()

Shiny UI module to add filter variable.

Usage

MatrixFilterStates$ui_add_filter_state(id, data)

Arguments

id

(character(1))
id of shiny module

data

(matrix)
data object for which to define a subset

Returns

shiny.tag

Method srv_add_filter_state()

Shiny server module to add filter variable

Module controls available choices to select as a filter variable. Selected filter variable is being removed from available choices. Removed filter variable gets back to available choices.

Usage

MatrixFilterStates$srv_add_filter_state(id, data, ...)

Arguments

id

(character(1))
shiny module instance id

data

(matrix)
data object for which to define a subset

...

ignored

Returns

moduleServer function which returns NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

MatrixFilterStates$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.