FilterState object for numeric variable

Manages choosing a numeric range

Super class

teal.slice::FilterState -> RangeFilterState

Methods

Public methods

RangeFilterState$new()

RangeFilterState$get_call()

RangeFilterState$get_keep_inf()

RangeFilterState$clone()

Inherited methods

teal.slice::FilterState$destroy_observers()

teal.slice::FilterState$format()

teal.slice::FilterState$get_state()

teal.slice::FilterState$print()

teal.slice::FilterState$server()

teal.slice::FilterState$set_state()

teal.slice::FilterState$ui()

Method new()

Initialize a FilterState object for range selection

Usage

RangeFilterState$new(
  x,
  x_reactive = reactive(NULL),
  extract_type = character(0),
  slice
)

Arguments

x

(numeric)
values of the variable used in filter

x_reactive

(reactive)
returning vector of the same type as x. Is used to update counts following the change in values of the filtered dataset. If it is set to reactive(NULL) then counts based on filtered dataset are not shown.

extract_type

(character(0), character(1))
whether condition calls should be prefixed by dataname. Possible values:

character(0) (default) varname in the condition call will not be prefixed

"list" varname in the condition call will be returned as <dataname>$<varname>

"matrix" varname in the condition call will be returned as <dataname>[, <varname>]

slice

(teal_slice)
object created using teal_slice(). teal_slice is stored in the class and set_state directly manipulates values within teal_slice. get_state returns teal_slice object which can be reused in other places. Beware, that teal_slice is a reactiveValues which means that changes in particular object are automatically reflected in all places which refer to the same teal_slice.

...

additional arguments to be saved as a list in private$extras field

Method get_call()

Returns reproducible condition call for current selection. For this class returned call looks like <varname> >= <min value> & <varname> <= <max value> with optional is.na(<varname>) and is.finite(<varname>).

Usage

RangeFilterState$get_call(dataname)

Arguments

dataname

name of data set; defaults to private$get_dataname()

Returns

(call)

Method get_keep_inf()

Returns current keep_inf selection

Usage

RangeFilterState$get_keep_inf()

Returns

(logical(1))

Method clone()

The objects of this class are cloneable with this method.

Usage

RangeFilterState$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

filter_state <- teal.slice:::RangeFilterState$new(
  x = c(NA, Inf, seq(1:10)),
  slice = teal_slice(varname = "x", dataname = "data")
)
shiny::isolate(filter_state$get_call())
#> NULL
filter_state$set_state(
  teal_slice(
    dataname = "data",
    varname = "x",
    selected = c(3L, 8L),
    keep_na = TRUE,
    keep_inf = TRUE
  )
)
shiny::isolate(filter_state$get_call())
#> is.na(x) | (is.infinite(x) | x >= 3 & x <= 8)

# working filter in an app
library(shiny)
library(shinyjs)

data_range <- c(runif(100, 0, 1), NA, Inf)
fs <- teal.slice:::RangeFilterState$new(
  x = data_range,
  slice = teal_slice(
    dataname = "data",
    varname = "x",
    selected = c(0.15, 0.93),
    keep_na = TRUE,
    keep_inf = TRUE
  )
)

ui <- fluidPage(
  useShinyjs(),
  teal.slice:::include_css_files(pattern = "filter-panel"),
  teal.slice:::include_js_files(pattern = "count-bar-labels"),
  column(4, div(
    h4("RangeFilterState"),
    fs$ui("fs")
  )),
  column(4, div(
    id = "outputs", # div id is needed for toggling the element
    h4("Condition (i.e. call)"), # display the subsetting call generated by this FilterState
    textOutput("condition_range"), br(),
    h4("Unformatted state"), # display raw filter state
    textOutput("unformatted_range"), br(),
    h4("Formatted state"), # display human readable filter state
    textOutput("formatted_range"), br()
  )),
  column(4, div(
    h4("Programmatic filter control"),
    actionButton("button1_range", "set drop NA", width = "100%"), br(),
    actionButton("button2_range", "set keep NA", width = "100%"), br(),
    actionButton("button3_range", "set drop Inf", width = "100%"), br(),
    actionButton("button4_range", "set keep Inf", width = "100%"), br(),
    actionButton("button5_range", "set a range", width = "100%"), br(),
    actionButton("button6_range", "set full range", width = "100%"), br(),
    actionButton("button0_range", "set initial state", width = "100%"), br()
  ))
)

server <- function(input, output, session) {
  fs$server("fs")
  output$condition_range <- renderPrint(fs$get_call())
  output$formatted_range <- renderText(fs$format())
  output$unformatted_range <- renderPrint(fs$get_state())
  # modify filter state programmatically
  observeEvent(
    input$button1_range,
    fs$set_state(teal_slice(dataname = "data", varname = "x", keep_na = FALSE))
  )
  observeEvent(
    input$button2_range,
    fs$set_state(teal_slice(dataname = "data", varname = "x", keep_na = TRUE))
  )
  observeEvent(
    input$button3_range,
    fs$set_state(teal_slice(dataname = "data", varname = "x", keep_inf = FALSE))
  )
  observeEvent(
    input$button4_range,
    fs$set_state(teal_slice(dataname = "data", varname = "x", keep_inf = TRUE))
  )
  observeEvent(
    input$button5_range,
    fs$set_state(
      teal_slice(dataname = "data", varname = "x", selected = c(0.2, 0.74))
    )
  )
  observeEvent(
    input$button6_range,
    fs$set_state(teal_slice(dataname = "data", varname = "x", selected = c(0, 1)))
  )
  observeEvent(
    input$button0_range,
    fs$set_state(
      teal_slice("data", "variable", selected = c(0.15, 0.93), keep_na = TRUE, keep_inf = TRUE)
    )
  )
}

if (interactive()) {
  shinyApp(ui, server)
}