Initializes FilteredDataset

Initializes FilteredDataset

Usage

init_filtered_dataset(
  dataset,
  dataname,
  keys = character(0),
  label = attr(dataset, "label"),
  metadata = NULL
)

Arguments

dataset

(data.frame or MultiAssayExperiment)

dataname

(character)
A given name for the dataset it may not contain spaces

keys

optional, (character)
Vector with primary keys

label

(character)
Label to describe the dataset

metadata

(named list or NULL)
Field containing metadata about the dataset. Each element of the list should be atomic and length one.

Note

Although this function is exported for use in other packages, it may be changed or removed in a future release at which point any code which relies on this exported function will need to be changed.

Examples

# DefaultFilteredDataset example
iris_fd <- teal.slice:::init_filtered_dataset(
  iris,
  dataname = "iris",
  metadata = list(type = "teal")
)
if (FALSE) {
shinyApp(
  ui = fluidPage(
    iris_fd$ui_add_filter_state(id = "add"),
    iris_fd$ui("dataset"),
    verbatimTextOutput("call"),
    verbatimTextOutput("metadata")
  ),
  server = function(input, output, session) {
    iris_fd$srv_add_filter_state(id = "add")
    iris_fd$server(id = "dataset")

    output$metadata <- renderText({
      paste("Type =", iris_fd$get_metadata()$type)
    })

    output$call <- renderText({
      paste(
        vapply(iris_fd$get_call(), deparse1, character(1), collapse = "\n"),
        collapse = "\n"
      )
    })
  }
)
}

# MAEFilteredDataset example
library(MultiAssayExperiment)
data(miniACC)
MAE_fd <- teal.slice:::init_filtered_dataset(miniACC, "MAE", metadata = list(type = "MAE"))
if (FALSE) {
shinyApp(
  ui = fluidPage(
    MAE_fd$ui_add_filter_state(id = "add"),
    MAE_fd$ui("dataset"),
    verbatimTextOutput("call"),
    verbatimTextOutput("metadata")
  ),
  server = function(input, output, session) {
    MAE_fd$srv_add_filter_state(id = "add")
    MAE_fd$server(id = "dataset")
    output$metadata <- renderText({
      paste("Type =", MAE_fd$get_metadata()$type)
    })
    output$call <- renderText({
      paste(
        vapply(MAE_fd$get_call(), deparse1, character(1), collapse = "\n"),
        collapse = "\n"
      )
    })
  }
)
}